Extract keyword/alias helpers into internals.kwargs and add _alias_kwargs
The internals package __init__ had grown into a grab-bag mixing keyword/alias
resolution with rc, location, and lazy-import helpers. Move the cohesive
keyword-argument cluster (_not_none, _alias_maps, the _pop_* poppers, and the
signature helpers) into a dedicated internals/kwargs.py, and re-export them from
the package so every existing 'from ..internals import _not_none' keeps working
unchanged. Add a new _alias_kwargs decorator that folds synonym keywords into
their canonical names with the same precedence and conflict warning as
_not_none, replacing the repetitive 'x = _not_none(x=x, y=y)' boilerplate at the
top of aliased functions.
